server/db: remove find and find_by from ActiveRecordMixin
