+++ server/tests/kit/db/models/mixins/test_active_record.py
@@ -1,12 +1,25 @@
+from typing import Any, Self
+
 import pytest
 
 from polar.kit.db.models.mixins import ActiveRecordMixin
+from polar.kit.extensions.sqlalchemy import sql
 from polar.postgres import AsyncSession
 from tests.fixtures.database import TestModel
 
 
 class ActiveRecord(TestModel, ActiveRecordMixin):
     ...
+
+    @classmethod
+    async def find_by(
+        cls,
+        session: AsyncSession,
+        **params: Any,
+    ) -> Self | None:
+        query = sql.select(cls).filter_by(**params)
+        res = await session.execute(query)
+        return res.scalars().unique().one_or_none()
 
 
 @pytest.mark.asyncio
@@ -17,39 +30,6 @@ async def test_create(session: AsyncSession) -> None:
     assert created.id is not None
     assert created.str_column == "Hello world"
     await created.delete(session)
-
-
-@pytest.mark.asyncio
-async def test_find(session: AsyncSession) -> None:
-    created = await ActiveRecord.create(session)
-    retrieved = await ActiveRecord.find(session, created.id)
-    assert retrieved is not None
-    assert retrieved.id == created.id
-    await created.delete(session)
-
-
-@pytest.mark.asyncio
-async def test_find_using_non_id(session: AsyncSession) -> None:
-    pseudo_id = 1337
-    created = await ActiveRecord.create(session, int_column=pseudo_id)
-
-    retrieved = await ActiveRecord.find(session, pseudo_id, key="int_column")
-    assert retrieved is not None
-    assert retrieved.id == created.id
-    assert retrieved.int_column == created.int_column
-    await retrieved.delete(session)
-
-
-@pytest.mark.asyncio
-async def test_find_by(session: AsyncSession) -> None:
-    pseudo_id = 1337
-    created = await ActiveRecord.create(session, int_column=pseudo_id)
-
-    retrieved = await ActiveRecord.find_by(session, id=created.id, int_column=pseudo_id)
-    assert retrieved is not None
-    assert retrieved.id == created.id
-    assert retrieved.int_column == created.int_column
-    await retrieved.delete(session)
 
 
 @pytest.mark.asyncio
@@ -97,18 +77,4 @@ async def test_autocommit_disabled_on_update(session: AsyncSession) -> None:
 
     await created.update(session, int_column=404, autocommit=False)
     retrieved = await ActiveRecord.find_by(session, int_column=404)
-    assert retrieved is None
-
-
-@pytest.mark.asyncio
-async def test_delete(session: AsyncSession) -> None:
-    created = await ActiveRecord.create(session)
-    assert created is not None
-    assert created.id
-
-    retrieved = await ActiveRecord.find(session, created.id)
-    assert retrieved is not None
-
-    await created.delete(session)
-    retrieved = await ActiveRecord.find(session, created.id)
     assert retrieved is None
